pyproject.toml
[tool.poetry]
name = "jwt-sample-server"
version = "0.1.0"
description = ""
authors = ["Naoya Yamashita <conao3@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.scripts]
jwt-sample-server = "jwt_sample_server.__main__:main"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
